requirements.txt
evernote3==1.25.14
beautifulsoup4==4.6.0
pydantic==1.8.2
python-dotenv==0.19.2
oauth2==1.9.0.post1
